Moved all rnn_components in separate folder
