+++ modules/cuda_plugin/src/ops/lstm_cell.hpp
@@ -7,8 +7,8 @@
 #include <cuda_operation_base.hpp>
 #include <ngraph/node.hpp>
 
-#include "rnn_components.hpp"
-#include "rnn_cudnn_components.hpp"
+#include "rnn_components/rnn_components.hpp"
+#include "rnn_components/rnn_cudnn_components.hpp"
 
 namespace CUDAPlugin {
