+++ modules/cuda_plugin/src/ops/rnn_components/lstm_sequence_components.cpp
@@ -5,15 +5,14 @@
 #include "lstm_sequence_components.hpp"
 
 #include <cuda/constant_factory.hpp>
+#include <cuda_op_buffers_extractor.hpp>
 #include <error.hpp>
 #include <gsl/gsl_assert>
 #include <ngraph/op/constant.hpp>
+#include <ngraph/op/reshape.hpp>
+#include <ngraph/op/squeeze.hpp>
+#include <ngraph/op/unsqueeze.hpp>
 #include <typeinfo>
-
-#include "cuda_op_buffers_extractor.hpp"
-#include "ngraph/op/reshape.hpp"
-#include "ngraph/op/squeeze.hpp"
-#include "ngraph/op/unsqueeze.hpp"
 
 namespace CUDAPlugin::RNN::Details {
