+++ modules/cuda_plugin/src/ops/rnn_components/lstm_sequence_cudnn_components.cpp
@@ -6,8 +6,7 @@
 
 #include <error.hpp>
 #include <gsl/gsl_assert>
-
-#include "converters.hpp"
+#include <ops/converters.hpp>
 
 namespace CUDAPlugin::RNN::Details {
