+++ modules/cuda_plugin/src/ops/rnn_components/rnn_sequence_components.cpp
@@ -1,3 +1,7 @@
+// Copyright (C) 2018-2021 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
 #include "rnn_sequence_components.hpp"
 
 #include <cuda/constant_factory.hpp>
